requirements.txt
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
fpdf2>=2.7.9
